select menu uses dialect
