+++ app/assets/javascripts/libfigure/define-figure.js
@@ -303,8 +303,7 @@
 /***/   });
 /***/   return ms;
 /***/ }
-/***/ 
-/***/ function moveTermsAndSubstitutionsForSelectMenu(dialect) {
+/***/ const moveTermsAndSubstitutionsForSelectMenu = (dialect) => {
 /***/   if (!dialect) { throw_up('must specify dialect to moveTermsAndSubstitutionsForSelectMenu'); }
 /***/   var mtas = moveTermsAndSubstitutions(dialect);
 /***/   var swing_index = mtas.findIndex(function (e) { return 'swing' === e.term;});
